requirements.txt
python-telegram-bot==20.6
openai==1.30.1
pytesseract
Pillow
python-docx
PyPDF2

python-dotenv